+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+### Fixed
+* Create correct cache key for downloads with an outSR
+
 ## [3.3.0] - 2016-03-24
 ### Changed
 * Simplify controller action for `getResource`